Split runtime morph logic
